feat: restructure app/theme package — zero standalone functions, Catalog-centric API
All standalone functions in app/theme are now methods on Theme or Catalog:
- Theme owns: Dump, clone, isCanonicalKey
- Catalog owns: all discovery, loading, installation, gallery access,
  and utility methods (parse, Load, Entries, InitBundled, Install, etc.)
- Unexport ThemeInfo → themeInfo, DefaultThemeName → defaultThemeName,
  GalleryNames → galleryNames, ColorKeys eliminated
- Update all callers in app/main.go, app/themes.go, and all test files
